requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
channels==4.1.0
channels-redis==4.2.0
python-dotenv==1.0.1
psycopg2-binary
drf-yasg
